This is full working Laravel Passport Backend. Part 11. User Todos Laravel API

+++ app/Http/Controllers/TodosController.php
@@ -14,7 +14,8 @@ class TodosController extends Controller
      */
     public function index()
     {
-        return Todo::all();
+    //    return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     /**
@@ -40,7 +41,12 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed,
+        ]);
+    //    $todo = Todo::create($data);
 
         return response($todo, 201);
     }
@@ -76,6 +82,10 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean',
@@ -93,7 +103,8 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
+    //    Todo::query()->update($data);
 
         return response('Updated', 200);
     }
@@ -106,6 +117,10 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
         return response('Deleted todo item', 200);
@@ -113,13 +128,34 @@ class TodosController extends Controller
 
     public function destroyCompleted(Request $request)
     {
+        // [4,7] todo ids we are passing in and want to delete
+        // [3,4,7] all of the users todo ids
+
+        // [7,10] validation would fail
+
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
+
         $request->validate([
             'todos' => 'required|array',
         ]);
 
-    //    Todo::destroy([1,2,3]);
         Todo::destroy($request->todos);
+    //    Todo::destroy([1,2,3]);
 
-        return response('Deleted', 200);
+        return response()->json('Deleted', 200);
+    //    return response('Deleted', 200);
     }
 }
